API/requirements.txt
Flask
mysql-connector
Pillow
opencv-python